pyproject.toml
[project]
name = "teuthology-metrics"
version = "0.1.0"
description = "Teuthology test run metrics"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "apscheduler>=3.11.1",
    "docopt-ng>=0.9.0",
    "drain3==0.9.11",
    "jinja2>=3.1.6",
    "opensearch-py>=3.1.0",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "ruff>=0.14.6",
]

[tool.ruff]
indent-width = 4
line-length = 79
exclude = [".venv"]

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
ignore = ["E402", "E741", "F522", "E203", "E731"]

[tool.ruff.lint.isort]
force-wrap-aliases = false
split-on-trailing-comma = true
